Revert "PM-439 Move provider-related selectors to integrations folder"
This reverts commit 4594c68dba4f0919a1000623b97fdcb9c5c514ad.
