+++ src/actions/blockchain.js
@@ -1,4 +1,5 @@
 import {
+  getGnosisConnection,
   initGnosisConnection,
   getCurrentBalance,
   getCurrentAccount,
@@ -14,10 +15,10 @@ import {
   getEtherTokens,
 } from 'api'
 
-import { timeoutCondition, getGnosisJsOptions } from 'utils/helpers'
+import { timeoutCondition, getGnosisJsOptions, weiToEth } from 'utils/helpers'
 import { GAS_COST } from 'utils/constants'
 import { createAction } from 'redux-actions'
-import { findDefaultProvider } from 'integrations/selectors'
+import { findDefaultProvider, getSelectedProvider } from 'selectors/blockchain'
 
 import { setActiveProvider } from './providers'
