+++ src/containers/DashboardPage/index.js
@@ -5,8 +5,7 @@ import DashboardPage from 'components/Dashboard'
 import { getMarkets, getAccountPredictiveAssets } from 'selectors/market'
 import { getAccountTrades } from 'selectors/marketTrades'
 import { getAccountShares } from 'selectors/marketShares'
-import { isGnosisInitialized } from 'selectors/blockchain'
-import { getCurrentAccount, getEtherTokensAmount, checkWalletConnection } from 'integrations/selectors'
+import { getCurrentAccount, getEtherTokensAmount, isGnosisInitialized, checkWalletConnection } from 'selectors/blockchain'
 import { requestMarkets, requestAccountTrades, requestAccountShares, redeemWinnings } from 'actions/market'
 import { requestGasPrice, requestEtherTokens } from 'actions/blockchain'
 import { weiToEth } from 'utils/helpers'
